+++ uCOS-III_Wrapper/uCOS-III/os_mutex.c
@@ -727,12 +727,6 @@ void  OSMutexPost (OS_MUTEX  *p_mutex,
         default:
             *p_err =  OS_ERR_OPT_INVALID;
              return;
-    }
-    if(opt != OS_OPT_POST_NONE)/*此opt选项只能为OS_OPT_POST_NONE*/
-    {
-        *p_err = OS_ERR_OPT_INVALID;
-        RT_DEBUG_LOG(OS_CFG_DBG_EN,("OSMutexPost: wrapper can't accept this option\n"));
-        return;
     }    
 #endif